Pull request #3331: Tenant id
Merge in SNORT/snort3 from ~RDEMPSTE/snort3:tenant_id to master

Squashed commit of the following:

commit 49bcaac681921062b79be6e17ffc319a9d7f831b
Author: Ron Dempster (rdempste) <rdempste@cisco.com>
Date:   Thu Apr 28 17:02:22 2022 -0400

    flow: change the padding and bits in the flow key to make it more clear

commit 76553e101331382ee5e7daca82fb34e513fbb23d
Author: Ron Dempster (rdempste) <rdempste@cisco.com>
Date:   Wed Apr 27 12:18:04 2022 -0400

    dce_rpc: update address space id in the smb keys

commit 3d78363477fcfe1c866ff62d73eb3a6a9970b3cf
Author: Ron Dempster (rdempste) <rdempste@cisco.com>
Date:   Tue Mar 22 12:04:08 2022 -0400

    flow: add inline cppcheck suppressions

commit f54d8a3cc078023a153b576e78583569dfd4bbb6
Author: Ron Dempster (rdempste) <rdempste@cisco.com>
Date:   Mon Mar 21 13:13:19 2022 -0400

    flow, network_inspectors, policy_selectors, stream: make address space id 32 bits and add a tenant id to the daq header

+++ src/flow/test/flow_control_test.cc
@@ -130,7 +130,7 @@ bool FlowKey::init(
     const SfIp*, uint16_t,
     const SfIp*, uint16_t,
     uint16_t, uint32_t,
-    uint16_t, int16_t, int16_t)
+    uint32_t, int16_t, int16_t)
 {
    return true;
 }
@@ -150,7 +150,7 @@ bool FlowKey::init(
     PktType, IpProtocol,
     const SfIp*, const SfIp*,
     uint32_t, uint16_t,
-    uint32_t, uint16_t, int16_t,
+    uint32_t, uint32_t, int16_t,
     int16_t)
 {
     return true;
